% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_qualtrics.R
\name{connect_to_qualtrics}
\alias{connect_to_qualtrics}
\title{connect_to_qualtrics}
\usage{
connect_to_qualtrics(
  qualtrics_credential,
  base_url = "riceuniversity.iad1.qualtrics.com",
  install = FALSE
)
}
\arguments{
\item{qualtrics_credential}{Add path to qualtrics credential file}

\item{base_url}{YOUR-QUALTRICS_BASE_URL - yourorganizationid.yourdatacenterid.qualtrics.com}

\item{install}{}
}
\value{
access to qualtrics API
}
\description{
connect_to_qualtrics
}
\examples{
connect_to_qualtrics(qualtrics_credential = "../qualtrics_cred.txt",
                                base_url = "riceuniversity.iad1.qualtrics.com",
                                install = FALSE)
}
