% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_survey_metadata_readme.R
\name{generate_survey_metadata_readme}
\alias{generate_survey_metadata_readme}
\title{generate_survey_metadata_readme}
\usage{
generate_survey_metadata_readme(
  qualtrics_df,
  width = "720px",
  height = "480px",
  remove_cols_str = "Name|DistributionChannel|Email|ExternalReference|IPAddress"
)
}
\arguments{
\item{qualtrics_df}{A Qualtrics dataframe}

\item{width}{Width of the scrollable box, can be in px unit}

\item{height}{Height of the scrollable box, can be in px unit}

\item{remove_cols_str}{String pattern of column names to remove}
}
\value{
html for scrollable table
}
\description{
Extracts questions and metadata tags from Qualtrics survey and creates a scrollable table
for friendly viewing for embedding in the survey README files.
}
\examples{
generate_survey_metadata_readme
}
