% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package.R
\name{get_package}
\alias{get_package}
\title{get_package}
\usage{
get_package(package)
}
\arguments{
\item{package}{A package name string}
}
\value{
access to package
}
\description{
This function checks whether a package is installed or not, if not installed, then installs the package and makes it available for the analysis. It will not install the package if it is already installed, then it will just make the package available readily.
}
\examples{
get_package(package_name)

}
