% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{get_survey_data}
\alias{get_survey_data}
\title{get_survey_data}
\usage{
get_survey_data(selected_kinetic_survey_kwd = "Kinetic-Demographic")
}
\arguments{
\item{selected_kinetic_survey_kwd}{}
}
\value{
Dataframe with metadata for all Kinetic Qualtrics surveys passed to the global environment as \code{kinetic_surveys}
Returns primarily dataframe for the desired survey from the Qualtrics API.
}
\description{
get_survey_data
}
\examples{
get_survey_data(selected_kinetic_survey_kwd = "Kinetic-Demographic")
}
