% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_packages_available.R
\name{make_packages_available}
\alias{make_packages_available}
\title{make_packages_available}
\usage{
make_packages_available(
  packages_needed = c("tidyverse", "qualtRics", "jsonlite", "kableExtra", "lubridate",
    "patchwork", "janitor", "roxygen2")
)
}
\arguments{
\item{packages_needed}{A character vector of package names}
}
\value{
packages
}
\description{
Enter a vector of packages to first check whether they are already installed.
If installed, then make them available by a library call.
If not installed, install each one and then make them available.
Uses the \code{get_package} function to install each package.
}
\examples{
packages_needed <- c("tidyverse", "lubridate")
make_packages_available(packages_needed)

}
