% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_qualtrics_df.R
\name{preprocess_qualtrics_df}
\alias{preprocess_qualtrics_df}
\title{preprocess_qualtrics_df}
\usage{
preprocess_qualtrics_df(qualtrics_df)
}
\arguments{
\item{qualtrics_df_name}{substitute this with the qualtrics dataframe name that you want to analyze}
}
\value{
Opens the dataframe viewer and returns a clean dataframe without the above fields and cases.
}
\description{
Read in desired Qualtrics dataframe from the environment and preprocess data
to remove minors, non-consents, identifying information, previews, test accounts,
incomplete responses, and uninformative columns.
}
\examples{
preprocess_qualtrics_df("./Kinetic_data.csv")
}
