% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics_cred.R
\name{read_qualtrics_cred}
\alias{read_qualtrics_cred}
\title{read_qualtrics_cred}
\usage{
read_qualtrics_cred(path_to_cred)
}
\arguments{
\item{path_to_cred}{qualtrics credential file name with path}
}
\value{
Qualtrics credential
}
\description{
This function reads in your qualtrics credential from the appropriate path.
The qualtrics credential is your Qualtrics API key that you can get from your Qualtrics account by generating key.
It is an alphanumeric string.
The credentials must be saved as a simple \code{.txt} files. Press enter at the end of the credential in the textfile to create a newline.
}
\examples{
read_qualtrics_cred(path_to_cred = "../Qualtrics/credentials/qualtrics_cred.txt")

}
